refactor: move types in GithubWorkflow to separated directories
