+++ Source/Application/Kysect.Shreks.Application.GithubWorkflow/Models/GithubSubmissionCreationResult.cs
@@ -1,5 +1,5 @@
 ﻿using Kysect.Shreks.Core.Submissions;
 
-namespace Kysect.Shreks.Application.GithubWorkflow;
+namespace Kysect.Shreks.Application.GithubWorkflow.Models;
 
 public record GithubSubmissionCreationResult(Submission Submission, bool IsCreated);
