+++ Source/Application/Kysect.Shreks.Application.GithubWorkflow/Submissions/IGithubSubmissionFactory.cs
@@ -1,0 +1,18 @@
+﻿using Kysect.Shreks.Application.Dto.Github;
+using Kysect.Shreks.Application.GithubWorkflow.Models;
+using Kysect.Shreks.Core.Submissions;
+
+namespace Kysect.Shreks.Application.GithubWorkflow.Submissions;
+
+public interface IGithubSubmissionFactory
+{
+    Task<GithubSubmissionCreationResult> CreateOrUpdateGithubSubmission(
+        GithubPullRequestDescriptor pullRequestDescriptor,
+        CancellationToken cancellationToken);
+
+    Task<GithubSubmission> CreateGithubSubmissionAsync(
+        Guid userId,
+        Guid assignmentId,
+        GithubPullRequestDescriptor pullRequestDescriptor,
+        CancellationToken cancellationToken);
+}
